+++ unstructured/utils.py
@@ -1,6 +1,7 @@
 import functools
 import importlib
 import json
+import logging
 import os
 import platform
 import subprocess
@@ -21,13 +22,33 @@ from typing import (
     cast,
 )
 
+import magic
+import matplotlib.pyplot as plt
 import requests
+from matplotlib import colors, patches
+from pdf2image import convert_from_path
+from PIL import Image
 from typing_extensions import ParamSpec
 
 from unstructured.__version__ import __version__
 
 DATE_FORMATS = ("%Y-%m-%d", "%Y-%m-%dT%H:%M:%S", "%Y-%m-%d+%H:%M:%S", "%Y-%m-%dT%H:%M:%S%z")
-
+TYPE_TO_COLOUR_MAP = {
+    "Text": "blue",
+    "FigureCaption": "orange",
+    "NarrativeText": "green",
+    "Title": "red",
+    "Address": "purple",
+    "EmailAddress": "brown",
+    "Image": "pink",
+    "PageBreak": "gray",
+    "Table": "olive",
+    "Header": "indigo",
+    "Footer": "coral",
+    "Formula": "gold",
+    "ListItem": "skyblue",
+    "UncategorizedText": "slateblue",
+}
 
 _T = TypeVar("_T")
 _P = ParamSpec("_P")
@@ -280,3 +301,105 @@ def scarf_analytics():
                 )
     except Exception:
         pass
+
+
+def draw_bboxes_on_pdf_or_image(
+    file_path,
+    elements,
+    desired_width=20,
+    save_images=False,
+    save_coordinates=False,
+    output_folder=None,
+    plot=False,
+):
+    """draw superimposed bounding boxes in pdf|images per page for the document in file_path"""
+
+    mimetype = magic.from_file(file_path)
+    if "PDF" in mimetype:
+        images = convert_from_path(file_path)
+    elif "image" in mimetype:
+        images = [Image.open(file_path)]
+    else:
+        raise TypeError(f"file mimetype should be PDF or image. Yours is {mimetype}")
+
+    bounding_boxes = [[] for _ in range(len(images))]
+    text_labels = [[] for _ in range(len(images))]
+    for ix, element in enumerate(elements):
+        n_page = element.metadata.page_number - 1
+        bounding_boxes[n_page].append(element.metadata.coordinates.to_dict()["points"])
+        text_labels[n_page].append(f"{ix}. {element.category}")
+
+    for page_ix, image in enumerate(images):
+        if desired_width:
+            aspect_ratio = images[0].width / images[0].height
+            desired_height = desired_width / aspect_ratio
+            fig, ax = plt.subplots(figsize=(desired_width, desired_height))
+        else:
+            fig, ax = plt.subplots()
+
+        ax.imshow(image)
+
+        for bbox, label in zip(bounding_boxes[page_ix], text_labels[page_ix]):
+            x_min, y_min = bbox[0]
+            x_max, y_max = bbox[2]
+            width = x_max - x_min
+            height = y_max - y_min
+            rect = patches.Rectangle(
+                (x_min, y_min),
+                width,
+                height,
+                linewidth=1,
+                edgecolor="black",
+                facecolor="none",
+            )
+            label_clean = "".join([ch for ch in label if ch.isalpha()]).strip()
+            rect.set_edgecolor(TYPE_TO_COLOUR_MAP[label_clean])
+            rect.set_facecolor(colors.to_rgba(TYPE_TO_COLOUR_MAP[label_clean], alpha=0.04))
+            ax.add_patch(rect)
+            ax.text(
+                x_min,
+                y_min - 5,
+                label,
+                fontsize=12,
+                weight="bold",
+                color=TYPE_TO_COLOUR_MAP[label_clean],
+                bbox={
+                    "facecolor": (1.0, 1.0, 1.0, 0.7),
+                    "edgecolor": (0.95, 0.95, 0.95, 0.0),
+                    "pad": 0.5,
+                },
+            )
+
+        if save_images or save_coordinates:
+            if not output_folder:
+                output_folder = "./"
+                logging.warning(
+                    "No output_folder defined. Storing predictions in relative path './'",
+                )
+
+            if save_images:
+                image_path = f"{output_folder}/images_with_bboxes"
+                if not os.path.exists(image_path):
+                    os.makedirs(image_path)
+                plt.savefig(f'{image_path}/{file_path.split("/")[-1]}_{page_ix}.png')
+
+            if save_coordinates:
+                annotations_path = f"{output_folder}/bboxes_coordinates"
+                if not os.path.exists(annotations_path):
+                    os.makedirs(annotations_path)
+                with open(
+                    f'{annotations_path}/{file_path.split("/")[-1]}_{page_ix}.json',
+                    "w",
+                ) as json_file:
+                    json.dump(
+                        [
+                            {e_save.category: e_save.metadata.coordinates.to_dict()["points"]}
+                            for e_save in elements
+                        ],
+                        json_file,
+                    )
+
+        if plot:
+            plt.show()
+        else:
+            plt.close()